+++ config/badge_presets.php
@@ -12,12 +12,16 @@ if (!defined('IN_PHPBB'))
 }
 
 return [
-	['key' => 'changelog', 'value' => 'changelog-v1.0.0-a808f3', 'title' => 'Changelog'],
-	['key' => 'custom',    'value' => 'SUBJECT-STATUS-COLOR',     'title' => 'Custom'],
-	['key' => 'license',   'value' => 'license-GPLv2-blue',       'title' => 'License GPLv2'],
-	['key' => 'v100',      'value' => 'version-1.0.0-FFFF00',     'title' => 'Version 1.0.0'],
-	['key' => 'v101',      'value' => 'version-1.0.1-f36808',     'title' => 'Version 1.0.1'],
-	['key' => 'v200',      'value' => 'version-2.0.0-ff5733',     'title' => 'Version 2.0.0'],
-	['key' => 'v300',      'value' => 'version-3.0.0-ff5733',     'title' => 'Version 3.0.0'],
-	['key' => 'v400',      'value' => 'version-4.0.0-ff5733',     'title' => 'Version 4.0.0'],
+	['key' => 'changelog', 'value' => 'changelog-v1.0.0-a808f3', 'lang_key' => 'BADGE_PRESET_CHANGELOG'],
+	['key' => 'custom',    'value' => 'SUBJECT-STATUS-COLOR',     'lang_key' => 'BADGE_PRESET_CUSTOM'],
+	['key' => 'license',   'value' => 'license-GPLv2-blue',       'lang_key' => 'BADGE_PRESET_LICENSE'],
+	['key' => 'v100',      'value' => 'version-1.0.0-FFFF00',     'lang_key' => 'BADGE_PRESET_V100'],
+	['key' => 'v101',      'value' => 'version-1.0.1-f36808',     'lang_key' => 'BADGE_PRESET_V101'],
+	['key' => 'v200',      'value' => 'version-2.0.0-ff5733',     'lang_key' => 'BADGE_PRESET_V200'],
+	['key' => 'v300',      'value' => 'version-3.0.0-ff5733',     'lang_key' => 'BADGE_PRESET_V300'],
+	['key' => 'v400',      'value' => 'version-4.0.0-ff5733',     'lang_key' => 'BADGE_PRESET_V400'],
+	['key' => 'phpbb33',   'value' => 'phpBB-3.3-blue',           'lang_key' => 'BADGE_PRESET_PHPBB33'],
+	['key' => 'stable',    'value' => 'stable-yes-brightgreen',   'lang_key' => 'BADGE_PRESET_STABLE'],
+	['key' => 'beta',      'value' => 'status-beta-yellow',       'lang_key' => 'BADGE_PRESET_BETA'],
+	['key' => 'build',     'value' => 'build-passing-brightgreen','lang_key' => 'BADGE_PRESET_BUILD'],
 ];